Fix reviewing assessment details
