--- app/components/assessment_details/previous_summaries_component.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-module AssessmentDetails
-  class PreviousSummariesComponent < ViewComponent::Base
-    def initialize(planning_application:, category:)
-      @planning_application = planning_application
-      @category = category
-    end
-
-    private
-
-    attr_reader :planning_application, :category
-
-    def assessment_details
-      planning_application.assessment_details.where(category:)[1..]
-    end
-  end
-end
