+++ app/components/reviewing/assessment_details/previous_summary_component.html.erb
