+++ app/views/planning_applications/review/assessment_details/_form.html.erb
@@ -49,7 +49,7 @@
         )
       ) %>
       <%= render(
-        AssessmentDetails::PreviousSummariesComponent.new(
+        Reviewing::AssessmentDetails::PreviousSummariesComponent.new(
           planning_application: planning_application,
           category: assessment_detail
         )
